+++ app/controllers/api/ads_manager_controller.rb
@@ -1,8 +1,8 @@
 class API::AdsManagerController < ApplicationController
 
   def index
-    lat = params[:lat].to_i
-    lon = params[:lon].to_i
+    lat = params[:lat].to_f
+    lon = params[:lon].to_f
     
     adlocation = AdsManager.where("lat between '#{lat+0.011111}' and '#{lat+0.444444}' and lon between '#{lon+0.000001}' and '#{lon+0.9999999}'")
     adlocation = {'adlocation' => adlocation}
